Rename applicationService.ts to application-service.ts
